Camera as SDL_FRect: float maths, round only the final draw position; nearest-filter textures
Render::camera becomes SDL_FRect instead of SDL_Rect -- SDL3's renderer
is float-native, so the old int camera forced a float -> int -> float
round-trip on every move and truncated sub-pixel motion. DrawTexture,
DrawRectangle, DrawLine and DrawCircle now do their camera-relative
maths entirely in float and round only the final destination once,
last (std::round on the combined camera + world position, not the
camera alone) -- rounding only the camera isn't enough, since entities
at non-integer positions would still land between pixels.

Window::scale and Render's local scale become float to match.

Textures::LoadSurface sets SDL_SCALEMODE_NEAREST right after creating
each texture, since SDL3 defaults to linear filtering, which blurs
pixel art and bleeds colour between adjacent tiles in a tileset.

Carried forward from L01_Simple_Engine. Squashed per the 2026-09
one-time history-rewrite waiver (course not yet started, nobody has
cloned): was 7f79cde + c126572.

+++ src/Window.h
@@ -25,7 +25,7 @@ public:
 	void GetWindowSize(int& width, int& height) const;
 
 	// Retrieve window scale
-	int GetScale() const;
+	float GetScale() const;
 
 public:
 	// The window we'll be rendering to
@@ -34,5 +34,5 @@ public:
 	std::string title;
 	int width = 1280;
 	int height = 720;
-	int scale = 1;
+	float scale = 1.0f;
 };
